Renamed CreatorData to CreatorSection to make it clearer that it is part of a form
